- add: next script 추가 블로 - fix: script 관리 컴포넌트로 분리
