+++ src/app/robots.ts
@@ -5,7 +5,6 @@ export default function robots(): MetadataRoute.Robots {
     rules: {
       userAgent: '*',
       allow: '/',
-      disallow: '/private/',
     },
     sitemap: 'https://jin-blog-blush.vercel.app/sitemap.xml',
   };
